feat(styles): Add home text anim to utils
